Constant-ify some magic numbers; more efficient decryption
Mostly these numbers relate to the layout of the APDU buffer, so they
can't be eliminated easily and are still kinda messy.

Also support more than 16 bytes at once in the decryption routine. Now
you can supply a maximum-length APDU (as long as the payload is a
multiple of 16 bytes).